==> 1. Preprocess.ipynb <==
import re

def remove_numbers_and_special_characters(unicode_line):
    """
    Replace all characters not in 'vietnamese_alphabet' (numbers and special characters) with ' ' (a blank space)
    
    Args:
        unicode_line: str
    Returns:
        str
    """
    vietnamese_alphabet = 'a-záàảãạâấầẩẫậăắằẳẵặđéèẻẽẹêếềểễệíìỉĩịóòỏõọôốồổỗộơớờởỡợúùủũụưứừửữựýỳỷỹỵ'
    pattern = '[^' + vietnamese_alphabet + ']'
    alphabet_only_string = re.sub(pattern, ' ', unicode_line)
    return ' '.join(alphabet_only_string.split())

def process_line(unicode_line):
    """
    1. Convert to lower case
    2. Remove newline (\n)
    3. Remove numbers and special character
     
    Args:
        unicode_line: str
    Returns:
        str
    """
    line = unicode_line.lower()
    line = line.strip('\n')
    return remove_numbers_and_special_characters(line)

with open("train.txt", "r", encoding='utf-8') as f_in, open("train_rm_special_chars.txt", "w", encoding='utf-8') as f_out:
    for line in f_in:
        processed_line = process_line(line)
        if processed_line: # check empty string
            f_out.write(processed_line + '\n')
