==> 2. Convert to training data.ipynb <==
import re
import os
from operator import itemgetter

def split_tone_mark_from_char(unicode_str):
    """
    0 = không có thanh/dấu
    1 = thanh sắc
    2 = thanh huyền
    3 = thanh hỏi
    4 = thanh ngã
    5 = thanh nặng
    6 = dấu mũ trong chữ â, ô, ê
    7 = dấu móc trong chữ ư, ơ
    8 = dấu trăng trong chữ ă
    9 = dấu gạch trong chữ đ
    
    Biến đổi chữ có thanh dấu -> chữ không dấu + (thanh, dấu)
    Ví dụ:
    â -> a, (0, 6)
    ặ -> a, (5, 8)
    ó -> o, (1, 0)
    đ -> d, (0, 9)
    c -> c, (0, 0)
    
    Args:
        unicode_str: str
    Returns:
        str_no_tone: str, string after removing tones
        tone_mark_tuples: list of tuples, each tuple represents tone marks of the corresponding character 
    """
    tone_marks = {
        1: "áấắéếíóốớúứý",
        2: "àầằèềìòồờùừỳ",
        3: "ảẩẳẻểỉỏổởủửỷ",
        4: "ãẫẵẽễĩõỗỡũữỹ",
        5: "ạậặẹệịọộợụựỵ",
        6: "âấầẩẫậôốồổỗộêếềểễệ",
        7: "ưứừửữựơớờởỡợ",
        8: "ăắằẳẵặ",
        9: "đ"
    }
    
    characters = {
        'a': "aáàảãạâấầẩẫậăắằẳẵặ",
        'd': "dđ",
        'e': "eéèẻẽẹêếềểễệ",
        'i': "iíìỉĩị",
        'o': "oóòỏõọôốồổỗộơớờởỡợ",
        'u': "uúùủũụưứừửữự",
        'y': "yýỳỷỹỵ"
    }
    
    str_no_tone = ""
    tone_mark_tuples = []
    for char in unicode_str:
        tone_mark_indices = [i for i, x in enumerate(tone_marks.values()) if x.find(char) != -1]
        if len(tone_mark_indices) == 0: # không có dấu thanh
            char_no_tone = char
            tone_mark_keys = (0, 0)
        else:
            char_no_tone_index = [i for i, x in enumerate(characters.values()) if x.find(char) != -1]
            char_no_tone = list(characters.keys())[char_no_tone_index[0]]
            if len(tone_mark_indices) == 1: # chỉ có dấu hoặc chỉ có thanh
                tone_mark_key = itemgetter(*tone_mark_indices)(list(tone_marks.keys()))
                tone_mark_keys = (tone_mark_key, 0) if tone_mark_key < 6 else (0, tone_mark_key)
            else: # có cả dấu lẫn thanh
                tone_mark_keys = itemgetter(*tone_mark_indices)(list(tone_marks.keys()))
                
        str_no_tone += char_no_tone
        tone_mark_tuples.append(tone_mark_keys)
    return str_no_tone, tone_mark_tuples

alphabet_dict = {
    'a': 0,
    'b': 1,
    'c': 2,
    'd': 3,
    'e': 4,
    'f': 5,
    'g': 6,
    'h': 7,
    'i': 8,
    'j': 9,
    'k': 10,
    'l': 11,
    'm': 12,
    'n': 13,
    'o': 14,
    'p': 15,
    'q': 16,
    'r': 17,
    's': 18,
    't': 19,
    'u': 20,
    'v': 21,
    'w': 22,
    'x': 23,
    'y': 24,
    'z': 25,
    ' ': 26
}

alphabet_size = len(alphabet_dict)

tone_dict = {
    (0, 0): 0,
    (0, 6): 1,
    (0, 7): 2,
    (0, 8): 3,
    (0, 9): 4,
    (1, 0): 5,
    (1, 6): 6,
    (1, 7): 7,
    (1, 8): 8,
    (2, 0): 9,
    (2, 6): 10,
    (2, 7): 11,
    (2, 8): 12,
    (3, 0): 13,
    (3, 6): 14,
    (3, 7): 15,
    (3, 8): 16,
    (4, 0): 17,
    (4, 6): 18,
    (4, 7): 19,
    (4, 8): 20,
    (5, 0): 21,
    (5, 6): 22,
    (5, 7): 23,
    (5, 8): 24
}

tone_size = len(tone_dict)

rm_tone_line, tone = split_tone_mark_from_char('cầu tăng cung không tăng khiến giá thuê văn phòng cũng trong xu hướng tăng theo')
print(rm_tone_line)
print(tone)

data_dir = 'data'
if not os.path.exists(data_dir):
    os.mkdir(data_dir)

with open("train_rm_special_chars.txt", "r", encoding='utf-8') as f_in:
    i = 0
    for line in f_in:
        line = line.strip('\n')
        line, tone = split_tone_mark_from_char(line)
        
        length = len(line)
        line_str = ' '.join(map(str,[alphabet_dict[char] for char in line])) # encode characters as integers
        tone_str = ' '.join(map(str,[tone_dict[t] for t in tone])) # encode tone marks as integers
        
        with open(os.path.join(data_dir, "length_" + str(length) + ".txt"), "a+", encoding='utf-8') as f_out:
            f_out.write(line_str + ',' + tone_str + '\n')
            
        i += 1
        if (i % 1000) == 0:
            print(i)
