==> 2b. Train-validation split.ipynb <==
import glob
import os
import ntpath
import random
import pandas as pd

validation_size=0.05
data_dir = 'data'
train_dir = 'training'
validation_dir = 'validation'

if not os.path.exists(train_dir):
    os.mkdir(train_dir)
if not os.path.exists(validation_dir):
    os.mkdir(validation_dir)

for i, file_path in enumerate(glob.glob(os.path.join(data_dir, '*.txt'))):
    file_name = ntpath.basename(file_path)
    df = pd.read_csv(file_path, header=None)
    
    n_samples = len(df)
    validation_idx = random.sample(range(n_samples), round(n_samples*validation_size))
    train_idx = list(set(range(n_samples)) - set(validation_idx))
    
    df_train = df.loc[train_idx]
    df_validation = df.loc[validation_idx]
    
    if len(df_train) > 0:
        df_train.to_csv(os.path.join(train_dir, file_name), header=None, index=False)
    if len(df_validation) > 0:
        df_validation.to_csv(os.path.join(validation_dir, file_name), header=None, index=False)
        
    if i % 100 == 0:
        print(i)
