==> 4. Process test file.ipynb <==
import pandas as pd
import csv

df = pd.read_csv('test_word_per_line.txt')

df['label'] = df['id'].str[:3]

label_list = list(df['label'].unique())

len(label_list)

test_dict = {}
for idx, label in enumerate(label_list):
    word_list = list(df[df['label'] == label]['no_tone'])
    word_list = list(map(str, word_list))
    sentence = ' '.join(word_list)
    test_dict[label] = sentence
    if idx % 100 == 0:
        print(idx)

csv_writer = csv.writer(open("test.csv", "w"))
csv_writer.writerow(["id", "sentence"])
for key, val in test_dict.items():
    csv_writer.writerow([key, val])

df_2 = pd.read_csv('test.csv')

df_2
